* [sfTrac #134] Added source build options in OpenFLUID Market * Updated Market tests

+++ src/openfluid/market/MarketSrcPackage.cpp
@@ -76,16 +76,6 @@ MarketSrcPackage::MarketSrcPackage(openfluid::base::FuncID_t ID, std::string Pac
 // =====================================================================
 
 
-void MarketSrcPackage::addBuildConfigOptions(const std::string& Options)
-{
-  m_BuildConfigOptions = Options;
-}
-
-
-// =====================================================================
-// =====================================================================
-
-
 void MarketSrcPackage::process()
 {
   if (!m_Initialized)
@@ -97,7 +87,7 @@ void MarketSrcPackage::process()
   if (m_CMakeCommand.empty())
     throw openfluid::base::OFException("OpenFLUID framework","MarketSrcPackage::process()","CMake command not defined");
 
-  if (!m_BuildConfigOptions.empty()) m_BuildConfigOptions = " " + m_BuildConfigOptions;
+  std::string BuildConfigOptions = composeFullBuildOptions(m_BuildConfigOptions);
 
   std::string BuildDir = m_TempBuildsDir + "/" + m_ID;
   std::string SrcInstallDir = m_MarketBagSrcDir + "/" + m_ID;
@@ -116,9 +106,9 @@ void MarketSrcPackage::process()
 
   std::string UntarCommand = m_CMakeCommand +" -E chdir " + SrcInstallDir+ " " + m_CMakeCommand + " -E tar xfz " + m_PackageDest;
 
-  std::string BuildConfigCommand = m_CMakeCommand +" -E chdir " + BuildDir+ " " + m_CMakeCommand + " " + SrcInstallDir + m_BuildConfigOptions;
+  std::string BuildConfigCommand = m_CMakeCommand +" -E chdir " + BuildDir+ " " + m_CMakeCommand + " " + SrcInstallDir + BuildConfigOptions;
 
-  std::string BuildCommand = m_CMakeCommand +" -E chdir " + BuildDir+ " " + m_CMakeCommand + " --build . --target "+m_ID;
+  std::string BuildCommand = m_CMakeCommand +" -E chdir " + BuildDir+ " " + m_CMakeCommand + " --build .";
 
   if (m_IsLogEnabled)
   {
@@ -141,6 +131,9 @@ void MarketSrcPackage::process()
   // building
   if (std::system(BuildCommand.c_str()) != 0)
     throw openfluid::base::OFException("OpenFLUID framework","MarketSrcPackage::process()","Error building package using CMake");
+
+  if (!boost::filesystem::exists(boost::filesystem::path(BuildDir+"/"+m_ID+openfluid::config::PLUGINS_EXT)))
+    throw openfluid::base::OFException("OpenFLUID framework","MarketSrcPackage::process()","Error finding built package");
 
   if (boost::filesystem::exists(boost::filesystem::path(m_MarketBagBinDir+"/"+m_ID+openfluid::config::PLUGINS_EXT)))
     boost::filesystem::remove(boost::filesystem::path(m_MarketBagBinDir+"/"+m_ID+openfluid::config::PLUGINS_EXT));
